# kbmk_backend-main/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'KBMK UNMUL API Documentation'
  description: 'Dokumentasi API untuk team Frontend'
  contact:
    email: admin@admin.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
paths:
...
  /schedules:
    get:
      tags:
        - Schedules
      summary: 'Menampilkan daftar semua jadwal'
      description: 'Mengembalikan array dari objek jadwal yang diurutkan berdasarkan tanggal'
      operationId: 5c29d50423d0d6b8699a2135c2fb5fb9
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
                type: object
    post:
      tags:
        - Schedules
      summary: 'Membuat jadwal baru'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: 7c01405ac6954959e3e88545b2405603
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        201:
          description: 'Jadwal berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Schedule created successfully'
                  data:
                    $ref: '#/components/schemas/Schedule'
                type: object
        422:
          description: 'Error validasi'
          content:
            application/json:
              schema:
                properties:
                  nama:
                    type: array
                    items:
                      type: string
                      example: 'The nama field is required.'
                  tanggal:
                    type: array
                    items:
                      type: string
                type: object
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  '/schedules/{id}':
    get:
      tags:
        - Schedules
      summary: 'Menampilkan detail jadwal berdasarkan ID'
      description: 'Mengembalikan satu objek jadwal'
      operationId: 37e90f6bc49dcc1ff99759cc62146615
      parameters:
        -
          name: id
          in: path
          description: 'ID dari jadwal'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Schedule'
                type: object
        404:
          description: 'Jadwal tidak ditemukan'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Schedule not found.'
                type: object
    put:
      tags:
        - Schedules
      summary: 'Memperbarui jadwal'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: 856d6d9f76c212ceb91643de33238158
      parameters:
        -
          name: id
          in: path
          description: 'ID dari jadwal'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                nama:
                  type: string
                  maxLength: 50
                tanggal:
                  type: string
                  format: date
                deskripsi:
                  type: string
              type: object
      responses:
        200:
          description: 'Jadwal berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Schedule updated successfully'
                  data:
                    $ref: '#/components/schemas/Schedule'
                type: object
        404:
          description: 'Schedule not found.'
        422:
          description: 'Error validasi.'
        401:
          description: Unauthorized.
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Schedules
      summary: 'Menghapus jadwal'
      description: "Memerlukan autentikasi dan role 'superadmin'"
      operationId: adcce9884f580db32f049a4d1ce6e71e
      parameters:
        -
          name: id
          in: path
          description: 'ID dari jadwal'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Jadwal berhasil dihapus'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Schedule deleted successfully.'
                type: object
        404:
          description: 'Schedule not found.'
        401:
          description: Unauthorized.
      security:
        -
          bearerAuth: []
components:
  schemas:
...
    Schedule:
      required:
        - nama
        - tanggal
        - deskripsi
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          maxLength: 50
          example: 'Rapat Rutinan'
        tanggal:
          type: string
          format: date
          example: '2024-12-25'
        deskripsi:
          type: string
          example: 'Membahas agenda kegiatan bulan depan.'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
...